functions/requirements.txt
firebase_functions~=0.1.0
firebase_admin~=6.0.0
google-generativeai>=0.4.0
python-dotenv==1.0.0
flask==2.3.2
flask-cors==4.0.0
